simplestreams: marshal "Size" as an integer
Size was previously a float64, as by default,
encoding/json unmarshals a number as a float64.

If a struct field's type is an integer, the
number will be un/marshalled as that integer
type. There's some reflection shenanigans in
the simplestreams package where type-specific
items are first unmarshalled into a value of
type map[string]interface{}, and then remarshalled
and unmarshalled into a map of type map[string]*T,
where T is the type-specific item.

The above approach poses a problem for numbers,
as it is first unmarshalled as a float64, then
remarshalled into scientific notation, and then
finally fails to unmarshal into the integer field
on account of being in scientific notation.

Now, instead, we have a new type ItemsMap which
has an underlying type of map[string]interface{},
and implements json.Unmarshaler. The method will
store the input []byte, which is later used to
unmarshal into the map[string]*T. Avoiding the
remarshal/unmarshal described previously avoids
the problem.

+++ environs/tools/simplestreams.go
@@ -8,6 +8,7 @@ package tools
 
 import (
 	"fmt"
+
 	"launchpad.net/juju-core/environs/simplestreams"
 )
 
@@ -44,13 +45,13 @@ func (tc *ToolsConstraint) Ids() ([]string, error) {
 
 // ToolsMetadata holds information about a particular tools tarball.
 type ToolsMetadata struct {
-	Release  string  `json:"release"`
-	Version  string  `json:"version"`
-	Arch     string  `json:"arch"`
-	Size     float64 `json:"size"`
-	Path     string  `json:"path"`
-	FileType string  `json:"ftype"`
-	SHA256   string  `json:"sha256"`
+	Release  string `json:"release"`
+	Version  string `json:"version"`
+	Arch     string `json:"arch"`
+	Size     int64  `json:"size"`
+	Path     string `json:"path"`
+	FileType string `json:"ftype"`
+	SHA256   string `json:"sha256"`
 }
 
 // Fetch returns a list of tools for the specified cloud matching the constraint.
